stage stuff
Former-commit-id: 100eb74fa45447566275863612d82463cbff9d80

+++ rllab/distributions/categorical.py
@@ -8,7 +8,7 @@ TINY = 1e-8
 def from_onehot_sym(x_var):
     ret = TT.zeros((x_var.shape[0],), x_var.dtype)
     nonzero_n, nonzero_a = TT.nonzero(x_var)[:2]
-    ret = TT.set_subtensor(ret[nonzero_n], nonzero_a)
+    ret = TT.set_subtensor(ret[nonzero_n], nonzero_a.astype('uint8'))
     return ret
